Fix small issues
 - fix the error message text in CActivbeMasternodeUpdateSentinelPing;
 - add empty string before public: in CActiveMasternode class declaration;
 - rename field sentinelPing in CMasternodePing to sentinelIsActual and change $
 - decrease sentinelVersion field size to uint16_t;

+++ src/activemasternode.h
@@ -42,8 +42,9 @@ private:
 
     //  sentinel ping data
     int64_t sentinelPing;
-    int sentinelVersion;
- public:
+    uint16_t sentinelVersion;
+
+public:
     // Keys for the active Masternode
     CPubKey pubKeyMasternode;
     CKey keyMasternode;
